[LOGMGR-298] Require a minimum of Java 11 for both compile and runtime. Remove the current MR JAR types and migrate them to the main source tree.
https://issues.redhat.com/browse/LOGMGR-298
