+++ .github/workflows/maven.yml
@@ -14,27 +14,18 @@ jobs:
 
     runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest]
         java: ['11', '17']
 
     steps:
-      - uses: actions/checkout@v2
-      - name: Set up JDK 8
-        uses: actions/setup-java@v2
-        with:
-          java-version: 8
-          distribution: 'temurin'
-          cache: 'maven'
+      - uses: actions/checkout@v3
       - name: Set up JDK ${{ matrix.java }}
         uses: actions/setup-java@v2
         with:
           java-version: ${{ matrix.java }}
           distribution: 'temurin'
           cache: 'maven'
-      - name: Build with Maven Linux
-        if: runner.os == 'Linux'
-        run: mvn -B clean verify -Djava8.home=$JAVA_HOME_8_X64
-      - name: Build with Maven Windows
-        if: runner.os == 'Windows'
-        run: mvn -B clean verify '-Djava8.home="%JAVA_HOME_8_X64%"'
+      - name: Build and Test with Java ${{ matrix.java }}
+        run: mvn -B clean verify
