--- core/src/test/java/org/jboss/logmanager/Environment.java
@@ -1,39 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- *
- * Copyright 2018 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.logmanager;
-
-/**
- * A helper for tests with specific environment information.
- *
- * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
- */
-public class Environment {
-
-    /**
-     * Checks whether or not this is a modular JVM.
-     *
-     * @return {@code true} if this is a modular JVM (Java 9+), otherwise {@code false}
-     *
-     * @see Jvm#isModular()
-     */
-    public static boolean isModularJvm() {
-        return Jvm.isModular();
-    }
-}
