+++ core/src/test/java/org/jboss/logmanager/SystemLoggerIT.java
