+++ core/src/test/java/org/jboss/logmanager/SystemLoggerMain.java
