+++ pom.xml
@@ -33,7 +33,7 @@
     <parent>
         <groupId>org.jboss</groupId>
         <artifactId>jboss-parent</artifactId>
-        <version>38</version>
+        <version>39</version>
     </parent>
 
     <licenses>
@@ -70,15 +70,10 @@
         <skipITs>${skipTests}</skipITs>
         <skipUTs>${skipTests}</skipUTs>
 
-        <!-- Please note this version override is only done to support the testCompile use-case for Java 9+ tests. The
-             testCompile does not seem to include test dependencies on the class path in 3.8.0-jboss-1 for some reason.
-        -->
-        <version.compiler.plugin>3.7.0-jboss-1</version.compiler.plugin>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
 
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
-
-        <jdk.min.version>9</jdk.min.version>
+        <jdk.min.version>11</jdk.min.version>
     </properties>
 
     <dependencyManagement>
@@ -194,36 +189,4 @@
             </plugin>
         </plugins>
     </build>
-
-    <profiles>
-        <!-- This profile is required to add the tools.jar to the class path for Java 8 for byteman -->
-        <profile>
-            <id>java8-test-tools</id>
-            <activation>
-                <jdk>[9,</jdk>
-                <property>
-                    <name>java8.home</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <id>java8-test</id>
-                                <phase>test</phase>
-                                <configuration>
-                                    <additionalClasspathElements>
-                                        <additionalClasspathElement>${java8.home}/lib/tools.jar</additionalClasspathElement>
-                                    </additionalClasspathElements>
-                                </configuration>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
 </project>
